requirements.txt
numpy
scipy
absl-py
tqdm